RESUME-RANKER
Resume Ranker (analyzer) is a web application designed to rank resumes based on their relevance to a given job description. The application uses TF-IDF vectorization and cosine similarity to perform the ranking. The project is built with Flask as the backend framework and uses HTML and CSS for the frontend.